Things spawn and player pew.

+++ src/common/macro.h
@@ -13,7 +13,7 @@
 #define MiB(bytes) (KiB(bytes) * 1024LL)
 #define GiB(bytes) (MiB(bytes) * 1024LL)
 
-#define SECONDS(usec) (usec * 1e6)
+#define SECONDS(usec) (u64)(usec * 1e6)
 
 #define MIN(x, y) (y ^ ((x ^ y) & -(x < y)))
 #define MAX(x, y) (x ^ ((x ^ y) & -(x < y)))
